apps/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
celery==5.3.4
redis==5.0.1
anthropic==0.18.1
openai==1.6.1
pgvector==0.2.4
pdfplumber==0.10.3
PyPDF2==3.0.1
ebooklib==0.18
openai-whisper==20231117
selenium==4.15.2
playwright==1.40.0
cryptography==41.0.7
garminconnect==0.2.19
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0